tests/wpt/tests/css/css-conditional/container-queries/grid-item-container.html

<!doctype html>
<title>CSS Container Queries Test: Grid item container</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#size-container">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  #grid {
    width: 300px;
    display: grid;
    grid-template-columns: 2fr 1fr;
  }
  .item {
    container-type: inline-size;
  }
  @container (width > 50px) {
    .item div { color: lime; }
  }
  @container (width > 150px) {
    .item div { color: green; }
  }
</style>
<div id="grid">
  <div class="item">
    <div id="target1"></div>
  </div>
  <div class="item">
    <div id="target2"></div>
  </div>
</div>
<script>
  setup(() => assert_implements_size_container_queries());

  test(() => {
    assert_equals(getComputedStyle(target1).color, "rgb(0, 128, 0)", "First item container should be 200px wide");
    assert_equals(getComputedStyle(target2).color, "rgb(0, 255, 0)", "Second item container should be 100px wide");
  }, "Check that children can query grid item containers");
</script>
